Revert C compatibility for files
