+++ src/mcu_program/common_src/lms7002m_filters.h
@@ -3,7 +3,15 @@
 
 #include "typedefs.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 uint8_t TuneTxFilter(const float_type bandwidth);
 uint8_t TuneRxFilter(const float_type rx_lpf_freq_RF);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif
